Move POJO stuff, scope exports
