+++ examples/apps/branching/src/container/getChangesFromHealthBot.ts
@@ -5,13 +5,43 @@
 
 import { NETWORK_askHealthBotForSuggestions } from "../healthBot.js";
 
-import {
-	diffGroceryListPOJO,
-	extractGroceryListPOJO,
-	type GroceryListChanges,
-	type GroceryListPOJO,
-	type IGroceryList,
-} from "./groceryList/index.js";
+import type { IGroceryList } from "./groceryList/index.js";
+import type {
+	GroceryListChanges,
+	GroceryListItemPOJO,
+	GroceryListPOJO,
+} from "./interfaces.js";
+
+const extractGroceryListPOJO = (groceryList: IGroceryList): GroceryListPOJO =>
+	groceryList.getItems();
+
+const diffGroceryListPOJO = (
+	baseGroceryListPOJO: GroceryListPOJO,
+	modifiedGroceryListPOJO: GroceryListPOJO,
+): GroceryListChanges => {
+	const removals: GroceryListItemPOJO[] = [];
+	for (const maybeRemoval of baseGroceryListPOJO) {
+		if (
+			!modifiedGroceryListPOJO.find(
+				(destinationItem) => destinationItem.id === maybeRemoval.id,
+			)
+		) {
+			removals.push(maybeRemoval);
+		}
+	}
+
+	const adds: GroceryListItemPOJO[] = [];
+	for (const maybeAdd of modifiedGroceryListPOJO) {
+		if (!baseGroceryListPOJO.find((sourceItem) => sourceItem.id === maybeAdd.id)) {
+			adds.push(maybeAdd);
+		}
+	}
+
+	return {
+		adds,
+		removals,
+	};
+};
 
 export const getChangesFromHealthBot = async (
 	groceryList: IGroceryList,
