+++ examples/apps/branching/src/container/groceryList/index.ts
@@ -5,14 +5,6 @@
 
 export { GroceryListFactory } from "./groceryList.js";
 export {
-	applyDiffToGroceryList,
-	diffGroceryListPOJO,
-	extractGroceryListPOJO,
-} from "./groceryListUtils.js";
-export {
-	GroceryListItemPOJO,
-	GroceryListPOJO,
-	GroceryListChanges,
 	IGroceryItem,
 	IGroceryList,
 	IGroceryListEvents,
