+++ examples/apps/branching/src/container/groceryList/interfaces.ts
@@ -5,7 +5,7 @@
 
 import type { IEvent, IEventProvider } from "@fluidframework/core-interfaces";
 
-// #region GroceryList interfaces
+// #region GroceryList
 /**
  * Interfaces for the GroceryList data object.
  */
@@ -31,21 +31,7 @@ export interface IGroceryList {
 	readonly disposed: boolean;
 }
 
-// #region Utils interfaces
-/**
- * Interfaces used for extracting, diffing, and applying changes to an IGroceryList.
- */
-
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
-export type GroceryListItemPOJO = { id: string; name: string };
-export type GroceryListPOJO = GroceryListItemPOJO[];
-// eslint-disable-next-line @typescript-eslint/consistent-type-definitions
-export type GroceryListChanges = {
-	adds: GroceryListItemPOJO[];
-	removals: GroceryListItemPOJO[];
-};
-
-// #region Runtime interfaces
+// #region Runtime
 /**
  * Interfaces that really should probably be in some runtime package instead.  Describe runtime objects
  * with scoped capabilities.
