+++ examples/apps/branching/src/container/index.ts
@@ -4,18 +4,9 @@
  */
 
 export {
-	GroceryListChanges,
-	GroceryListItemPOJO,
-	GroceryListPOJO,
-	IGroceryItem,
-	IGroceryList,
-	IGroceryListEvents,
-} from "./groceryList/index.js";
-export {
 	ISuggestionGroceryItem,
 	ISuggestionGroceryList,
 	ISuggestionGroceryListEvents,
 	SuggestionState,
 } from "./interfaces.js";
-export { SuggestionGroceryList } from "./suggestionGroceryList.js";
 export { GroceryListContainerRuntimeFactory } from "./runtimeFactory.js";
